backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.26.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1
pydantic[email]==2.6.1
python-decouple==3.8
websockets==12.0
aioredis==2.0.1
httpx==0.26.0
pillow==10.2.0
openpyxl==3.1.2
reportlab==4.0.9
jinja2==3.1.3
pydantic-settings==2.1.0
asyncpg==0.29.0
twilio==8.13.0
aiogram==3.4.1
stripe==8.2.0
click==8.1.7
cryptography==42.0.2
aiofiles==23.2.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
selenium==4.17.2
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0
flower==2.0.1